+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "cph-ng" extension will be documented in this file.
 
+## [Unreleased]
+
+- **feat**: add stderrThreshold configuration to control stderr output handling
+
 ## 0.0.11
 
 - **feat**: import and export from embedded data